Rename containers to pages in i18n
