+++ packages/plugins/i18n/admin/src/components/Initializer/index.js
@@ -6,8 +6,8 @@
 
 import { useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
-import pluginId from '../pluginId';
-import useLocales from '../hooks/useLocales';
+import pluginId from '../../pluginId';
+import useLocales from '../../hooks/useLocales';
 
 const Initializer = ({ setPlugin }) => {
   const { isLoading, locales } = useLocales();
