+++ packages/plugins/i18n/admin/src/index.js
@@ -5,8 +5,8 @@ import pluginPkg from '../../package.json';
 import pluginLogo from './assets/images/logo.svg';
 import CheckboxConfirmation from './components/CheckboxConfirmation';
 import CMEditViewInjectedComponents from './components/CMEditViewInjectedComponents';
-import Initializer from './containers/Initializer';
-import SettingsPage from './containers/SettingsPage';
+import Initializer from './components/Initializer';
+import SettingsPage from './pages/SettingsPage';
 import LocalePicker from './components/LocalePicker';
 import middlewares from './middlewares';
 import pluginPermissions from './permissions';
