+++ packages/plugins/i18n/admin/src/pages/SettingsPage/tests/SettingsPage.test.js
